+++ NEWS.md
@@ -14,6 +14,8 @@
 
 * Fix a problem with `t1kable` when the row with the label contains data.
 
+* Choose between `qflextable`, `flextable` and `regulartable` in `t1flex`.
+
 # table1 1.4.2
 
 * Add option `rounding.fn` to `stats.apply.rounding()` to allow changing the